Move test.factories to pootle_pytest.factories
