+++ tests/models/project.py
@@ -11,7 +11,7 @@ import pytest
 
 from pootle_project.models import Project
 
-from ..factories import UserFactory
+from pootle_pytest.factories import UserFactory
 from ..fixtures.models.permission_set import _require_permission_set
 
 from pootle_pytest.utils import items_equal
